Fix `simple_case_expression`
```
case <expression> when <condition> then ...
```

changed to

```
case <expression> when <expression> then ...
                       ^^^^^^^^^^^
```

Due to the spec: "In a simple `CASE` expression, Oracle Database searches for the first `WHEN ... THEN` pair for which `expr` *is equal to* `comparison_expr`* and returns `return_expr`."

I.e. the comparison for equality is implicit.